+++ src/main/java/fi/evident/raudikko/SpellingSuggester.java
@@ -30,65 +30,19 @@
  * the terms of any one of the MPL, the GPL or the LGPL.
  */
 
-package fi.evident.raudikko.internal.utils;
+package fi.evident.raudikko;
 
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
-import static java.util.Collections.unmodifiableCollection;
+/**
+ * Provides suggestions for misspelled words.
+ */
+public interface SpellingSuggester {
 
-public final class CharMap<T> {
-
-    private final Object[] low = new Object[256];
-    private @Nullable Map<Character, T> high = null;
-
-    public void put(char key, T value) {
-        if (key < low.length)
-            low[key] = value;
-        else {
-            if (high == null)
-                high = new HashMap<>();
-
-            high.put(key, value);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    public @Nullable T get(char key) {
-        if (key < low.length)
-            return (T) low[key];
-        else if (high != null)
-            return high.get(key);
-        else
-            return null;
-    }
-
-    public @NotNull T getOrDefault(char key, @NotNull T defaultValue) {
-        var result = get(key);
-        return result != null ? result : defaultValue;
-    }
-
-    public @NotNull Collection<Character> keys() {
-        var result = new ArrayList<Character>();
-        for (int i = 0; i < low.length; i++)
-            if (low[i] != null)
-                result.add((char) i);
-
-        if (high != null)
-            result.addAll(high.keySet());
-
-        return unmodifiableCollection(result);
-    }
-
-    public @NotNull CharMap<T> copy() {
-        var result = new CharMap<T>();
-        System.arraycopy(low, 0, result.low, 0, low.length);
-        result.high = high != null ? new HashMap<>(high) : null;
-        return result;
-    }
+    /**
+     * Given a word, provides a list of spelling suggestions for it.
+     */
+    @NotNull List<String> provideSpellingSuggestions(@NotNull String word);
 }